+++ tests/integration/test_auth.py
@@ -880,3 +880,35 @@ class TestCSRFErrorHandler:
             )
             assert response.status_code == 200
             assert b"session has expired" in response.data or b"log in again" in response.data
+
+
+class TestAccountDeletion:
+    """Test account deletion functionality."""
+
+    def test_delete_account_page_requires_login(self, client):
+        """Test that delete account page requires login."""
+        response = client.get("/delete_account")
+        assert response.status_code == 302  # Redirect to login
+
+    def test_delete_account_soft_delete_preserves_user_data(self, client, app, auth_user):
+        """Test that account deletion uses soft delete to preserve user data for history."""
+        with app.app_context():
+            user = auth_user()
+            user_id = user.id
+            user_email = user.email
+
+            login_user(client, user.email)
+
+            response = client.post(
+                "/delete_account", data={"confirmation": "DELETE MY ACCOUNT"}, follow_redirects=True
+            )
+
+            assert response.status_code == 200
+
+            # Verify user was soft deleted (not hard deleted)
+            soft_deleted_user = db.session.get(User, user_id)
+            assert soft_deleted_user is not None
+            assert soft_deleted_user.is_deleted is True
+            assert soft_deleted_user.deleted_at is not None
+            assert "deleted_" in soft_deleted_user.email  # Email should be anonymized
+            assert soft_deleted_user.email != user_email  # Email changed
